requirements.txt
# Core dependencies
asyncio
typing
enum
dataclasses

# External dependencies
bcrypt>=4.0.0
PyJWT>=2.6.0
aiohttp>=3.8.0
coverage>=7.0.0
numpy>=1.21.0
sentence-transformers>=2.2.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0